Added access and error handling

+++ central_branch/views.py
@@ -3645,48 +3645,58 @@ def event_feedback(request, event_id):
 @member_login_permission
 def event_google_calendar(request, event_id):
 
-    current_user=renderData.LoggedinUser(request.user) #Creating an Object of logged in user with current users credentials
-    user_data=current_user.getUserData() #getting user data as dictionary file
-    sc_ag=PortData.get_all_sc_ag(request=request)
+    try:
+        current_user=renderData.LoggedinUser(request.user) #Creating an Object of logged in user with current users credentials
+        user_data=current_user.getUserData() #getting user data as dictionary file
+        sc_ag=PortData.get_all_sc_ag(request=request)
 
-    if(request.method == "POST"):
-        if('update_event_gc' in request.POST):
-            google_calendar_publish_event_status = request.POST.get('publish_event_gc')
-            attendeeOption = request.POST.get('attendeeList')
-            event_description_for_gc = request.POST.get('event_description_for_gc')
-            add_attendee_names = request.POST.getlist('attendee_name')
-            add_attendee_emails = request.POST.getlist('attendee_email')
+        has_access = Branch_View_Access.get_event_edit_access(request)
+        if has_access:
+            if(request.method == "POST"):
+                if('update_event_gc' in request.POST):
+                    google_calendar_publish_event_status = request.POST.get('publish_event_gc')
+                    attendeeOption = request.POST.get('attendeeList')
+                    event_description_for_gc = request.POST.get('event_description_for_gc')
+                    add_attendee_names = request.POST.getlist('attendee_name')
+                    add_attendee_emails = request.POST.getlist('attendee_email')
 
-            documents = None
-            if request.FILES.get('document'):
-                documents = request.FILES.getlist('document')
+                    documents = None
+                    if request.FILES.get('document'):
+                        documents = request.FILES.getlist('document')
 
-            publish_event_gc = Branch.button_status(google_calendar_publish_event_status)
-            Branch.update_event_google_calendar(request=request, event_id=event_id, description=event_description_for_gc, publish_event_gc=publish_event_gc, attendeeOption=attendeeOption, add_attendee_names=add_attendee_names, add_attendee_emails=add_attendee_emails, documents=documents)
-        if('remove_attachment') in request.POST:
-            attachment_id = request.POST.get('remove_attachment')
-            Branch.delete_attachment(request, attachment_id)
+                    publish_event_gc = Branch.button_status(google_calendar_publish_event_status)
+                    Branch.update_event_google_calendar(request=request, event_id=event_id, description=event_description_for_gc, publish_event_gc=publish_event_gc, attendeeOption=attendeeOption, add_attendee_names=add_attendee_names, add_attendee_emails=add_attendee_emails, documents=documents)
+                if('remove_attachment') in request.POST:
+                    attachment_id = request.POST.get('remove_attachment')
+                    Branch.delete_attachment(request, attachment_id)
 
-    event = Events.objects.get(id=event_id)
-    event_gc_attachments = Google_Calendar_Attachments.objects.filter(event_id=event)
-    form = EventFormGC({'event_description_for_gc' : event.event_description_for_gc})
-    is_event_published_gc = event.publish_in_google_calendar
-    additional_attendees = event.additional_attendees
+            event = Events.objects.get(id=event_id)
+            event_gc_attachments = Google_Calendar_Attachments.objects.filter(event_id=event)
+            form = EventFormGC({'event_description_for_gc' : event.event_description_for_gc})
+            is_event_published_gc = event.publish_in_google_calendar
+            additional_attendees = event.additional_attendees
+            
+
+            context = {
+                'is_branch':True,
+                'user_data':user_data,
+                'all_sc_ag':sc_ag,
+                'event':event,
+                'is_event_published_gc':is_event_published_gc,
+                'event_id':event_id,
+                'form':form,
+                'event_gc_attachments':event_gc_attachments,
+                'additional_attendees':additional_attendees
+            }
+
+            return render(request, 'Events/event_edit_google_calendar.html', context)
+        else:
+            return render(request,'access_denied2.html', {'all_sc_ag':sc_ag,'user_data':user_data,})
     
-
-    context = {
-        'is_branch':True,
-        'user_data':user_data,
-        'all_sc_ag':sc_ag,
-        'event':event,
-        'is_event_published_gc':is_event_published_gc,
-        'event_id':event_id,
-        'form':form,
-        'event_gc_attachments':event_gc_attachments,
-        'additional_attendees':additional_attendees
-    }
-
-    return render(request, 'Events/event_edit_google_calendar.html', context)
+    except Exception as e:
+        logger.error("An error occurred at {datetime}".format(datetime=datetime.now()), exc_info=True)
+        ErrorHandling.saveSystemErrors(error_name=e,error_traceback=traceback.format_exc())
+        return custom_500(request)
 
 @login_required
 @member_login_permission
